+++ packages/expo-cli/package.json
@@ -100,7 +100,6 @@
     "got": "^11.1.4",
     "hashids": "1.1.4",
     "indent-string": "4.0.0",
-    "is-root": "^2.1.0",
     "js-yaml": "^3.13.1",
     "keychain": "1.3.0",
     "leven": "^3.1.0",
