+++ packages/xdl/package.json
@@ -60,6 +60,7 @@
     "indent-string": "3.2.0",
     "internal-ip": "4.3.0",
     "is-reachable": "^4.0.0",
+    "is-root": "^2.1.0",
     "json-schema-deref-sync": "^0.13.0",
     "latest-version": "5.1.0",
     "lodash": "^4.17.19",
